[tune] add TrialExecutor
