+++ python/ray/tune/trial.py
@@ -21,7 +21,6 @@ from ray.tune.result import (DEFAULT_RESULTS_DIR, DONE, HOSTNAME, PID,
 from ray.utils import random_string, binary_to_hex
 
 DEBUG_PRINT_INTERVAL = 5
-MAX_LEN_IDENTIFIER = 130
 
 
 def date_str():
@@ -119,7 +118,6 @@ class Trial(object):
         self.last_result = None
         self._checkpoint_path = restore_path
         self._checkpoint_obj = None
-        self.runner = None
         self.status = Trial.PENDING
         self.location = None
         self.logdir = None
@@ -136,97 +134,6 @@ class Trial(object):
     def generate_id(cls):
         return binary_to_hex(random_string())[:8]
 
-    def start(self, checkpoint_obj=None):
-        """Starts this trial.
-
-        If an error is encountered when starting the trial, an exception will
-        be thrown.
-
-        Args:
-            checkpoint_obj (obj): Optional checkpoint to resume from.
-        """
-
-        self._setup_runner()
-        if checkpoint_obj:
-            self.restore_from_obj(checkpoint_obj)
-        elif self._checkpoint_path:
-            self.restore_from_path(self._checkpoint_path)
-        elif self._checkpoint_obj:
-            self.restore_from_obj(self._checkpoint_obj)
-
-    def stop(self, error=False, error_msg=None, stop_logger=True):
-        """Stops this trial.
-
-        Stops this trial, releasing all allocating resources. If stopping the
-        trial fails, the run will be marked as terminated in error, but no
-        exception will be thrown.
-
-        Args:
-            error (bool): Whether to mark this trial as terminated in error.
-            error_msg (str): Optional error message.
-            stop_logger (bool): Whether to shut down the trial logger.
-        """
-
-        if error:
-            self.status = Trial.ERROR
-        else:
-            self.status = Trial.TERMINATED
-
-        try:
-            if error_msg and self.logdir:
-                self.num_failures += 1
-                error_file = os.path.join(self.logdir,
-                                          "error_{}.txt".format(date_str()))
-                with open(error_file, "w") as f:
-                    f.write(error_msg)
-                self.error_file = error_file
-            if self.runner:
-                stop_tasks = []
-                stop_tasks.append(self.runner.stop.remote())
-                stop_tasks.append(self.runner.__ray_terminate__.remote())
-                # TODO(ekl)  seems like wait hangs when killing actors
-                _, unfinished = ray.wait(
-                    stop_tasks, num_returns=2, timeout=250)
-        except Exception:
-            print("Error stopping runner:", traceback.format_exc())
-            self.status = Trial.ERROR
-        finally:
-            self.runner = None
-
-        if stop_logger and self.result_logger:
-            self.result_logger.close()
-            self.result_logger = None
-
-    def pause(self):
-        """We want to release resources (specifically GPUs) when pausing an
-        experiment. This results in a state similar to TERMINATED."""
-
-        assert self.status == Trial.RUNNING, self.status
-        try:
-            self.checkpoint(to_object_store=True)
-            self.stop(stop_logger=False)
-            self.status = Trial.PAUSED
-        except Exception:
-            print("Error pausing runner:", traceback.format_exc())
-            self.status = Trial.ERROR
-
-    def unpause(self):
-        """Sets PAUSED trial to pending to allow scheduler to start."""
-        assert self.status == Trial.PAUSED, self.status
-        self.status = Trial.PENDING
-
-    def resume(self):
-        """Resume PAUSED trials. This is a blocking call."""
-
-        assert self.status == Trial.PAUSED, self.status
-        self.start()
-
-    def train_remote(self):
-        """Returns Ray future for one iteration of training."""
-
-        assert self.status == Trial.RUNNING, self.status
-        return self.runner.train.remote()
-
     def should_stop(self, result):
         """Whether the given result meets this trial's stopping criteria."""
 
@@ -240,6 +147,7 @@ class Trial(object):
                 return True
 
         return False
+
 
     def should_checkpoint(self):
         """Whether this trial is due for checkpointing."""
@@ -294,6 +202,20 @@ class Trial(object):
         return self._checkpoint_path is not None or \
             self._checkpoint_obj is not None
 
+    def save_to_object(self):
+        """save the state of this trial to object
+
+        :return: object to store the state.
+        """
+        return None
+
+    def save_to_path(self):
+        """ save the state of trial to path
+
+        :return: path to store the state.
+        """
+        return None
+
     def checkpoint(self, to_object_store=False):
         """Checkpoints the state of this trial.
 
@@ -301,47 +223,19 @@ class Trial(object):
             to_object_store (bool): Whether to save to the Ray object store
                 (async) vs a path on local disk (sync).
         """
-
         obj = None
         path = None
         if to_object_store:
-            obj = self.runner.save_to_object.remote()
+            obj = self.save_to_object()
         else:
-            path = ray.get(self.runner.save.remote())
+            path = self.save_to_path()
         self._checkpoint_path = path
         self._checkpoint_obj = obj
 
         if self.verbose:
             print("Saved checkpoint for {} to {}".format(self, path or obj))
+
         return path or obj
-
-    def restore_from_path(self, path):
-        """Restores runner state from specified path.
-
-        Args:
-            path (str): A path where state will be restored.
-        """
-
-        if self.runner is None:
-            print("Unable to restore - no runner")
-        else:
-            try:
-                ray.get(self.runner.restore.remote(path))
-            except Exception:
-                print("Error restoring runner:", traceback.format_exc())
-                self.status = Trial.ERROR
-
-    def restore_from_obj(self, obj):
-        """Restores runner state from the specified object."""
-
-        if self.runner is None:
-            print("Unable to restore - no runner")
-        else:
-            try:
-                ray.get(self.runner.restore_from_object.remote(obj))
-            except Exception:
-                print("Error restoring runner:", traceback.format_exc())
-                self.status = Trial.ERROR
 
     def update_last_result(self, result, terminate=False):
         if terminate:
@@ -353,34 +247,6 @@ class Trial(object):
             self.last_debug = time.time()
         self.last_result = result
         self.result_logger.on_result(self.last_result)
-
-    def _setup_runner(self):
-        self.status = Trial.RUNNING
-        cls = ray.remote(
-            num_cpus=self.resources.cpu,
-            num_gpus=self.resources.gpu)(self._get_trainable_cls())
-        if not self.result_logger:
-            if not os.path.exists(self.local_dir):
-                os.makedirs(self.local_dir)
-            self.logdir = tempfile.mkdtemp(
-                prefix="{}_{}".format(
-                    str(self)[:MAX_LEN_IDENTIFIER], date_str()),
-                dir=self.local_dir)
-            self.result_logger = UnifiedLogger(self.config, self.logdir,
-                                               self.upload_dir)
-        remote_logdir = self.logdir
-
-        def logger_creator(config):
-            # Set the working dir in the remote process, for user file writes
-            if not os.path.exists(remote_logdir):
-                os.makedirs(remote_logdir)
-            os.chdir(remote_logdir)
-            return NoopLogger(config, remote_logdir)
-
-        # Logging for trials is handled centrally by TrialRunner, so
-        # configure the remote runner to use a noop-logger.
-        self.runner = cls.remote(
-            config=self.config, logger_creator=logger_creator)
 
     def _get_trainable_cls(self):
         return ray.tune.registry._global_registry.get(
@@ -405,3 +271,19 @@ class Trial(object):
         if self.experiment_tag:
             identifier += "_" + self.experiment_tag
         return identifier
+
+    class Holder():
+        IMPL_CLS = None # refer to the subclass of Trial.
+
+    def __new__(cls, *args, **kwargs):
+        return object.__new__(Trial.Holder.IMPL_CLS)
+
+def register_trial_impl(cls):
+    """register Trial type
+
+    any call to Trial() create a cls object
+
+    Args:
+        cls: cls should be a subclass of Trial
+    """
+    Trial.Holder.IMPL_CLS = cls
